requirements.txt
fastapi
uvicorn
requests
python-dotenv
streamlit
langchain
langchain-google-genai
faiss-cpu
